v0.25.01 (15 Weeks Left)
-Added AtkSpd,Range and MvmentSpd Crystals. completing the all the stat consumables

*items have yet to be coded.
*Full cycle still needs to be worked on.

+++ Cattler/Assets/1.MainProj/Script/SFX.cs
@@ -32,7 +32,7 @@ public class StatFX : MonoBehaviour
         }
     }
 
-    public void SetText(int amount, string stat)
+    public void SetText(float amount, string stat)
     {
         text.text = $"+{amount} {stat}"; 
     }
